chore: remove unnecessary files
