--- packages/sysweb3-keyring/src/trezor/init.ts
@@ -1,12 +1,0 @@
-import TrezorConnect from 'trezor-connect';
-
-export const initialize = async () => {
-  const trezorConnection = await TrezorConnect.init({
-    manifest: {
-      appUrl: 'https://paliwallet.com/',
-      email: 'pali@pollum.io',
-    },
-  });
-
-  return trezorConnection;
-};
